sentiment_analysis_project/requirements.txt
annotated-types==0.7.0
anyio==4.6.0
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
distro==1.9.0
Django==5.1.1
djangorestframework==3.15.2
et-xmlfile==1.1.0
exceptiongroup==1.2.2
groq==0.11.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
numpy==2.1.1
openpyxl==3.1.5
pandas==2.2.3
pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
